Move ElmahResult tests to own file.
